use float iso double in critcal path, is precise enough and gives a small speedup
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@6438 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/sound/Resample.hh
@@ -25,10 +25,10 @@ private:
 	void calcOutput(FilterIndex startFilterIndex, float* output);
 	void prepareData(unsigned extra);
 
-	double ratio;
-	double floatIncr;
-	double normFactor;
-	double lastPos;
+	float ratio;
+	float floatIncr;
+	float normFactor;
+	float lastPos;
 	FilterIndex increment;
 	unsigned halfFilterLen;
 	unsigned bufCurrent;
